refactor(streaming): move the kinesis producer out of the service directory
